extract generator rust
