+++ frb_codegen/src/generator/rust/ty.rs
@@ -16,5 +16,29 @@ pub enum TypeGenerator {
 
 #[enum_dispatch]
 pub trait TypeRustGeneratorTrait {
-    // TODO
+    fn wire2api_body(&self) -> String;
+
+    fn wire_struct_fields(&self) -> Vec<String> {
+        vec![]
+    }
+
+    fn structs(&self) -> String {
+        "".to_string()
+    }
+
+    fn allocate_funcs(&self, collector: &ExternFuncCollector) -> String {
+        "".to_string()
+    }
+
+    fn impl_intodart(&self) -> String {
+        "".to_string()
+    }
+
+    fn new_with_nullptr(&self) -> String {
+        "".to_string()
+    }
+
+    fn imports(&self) -> Option<String> {
+        "".to_string()
+    }
 }
