+++ frb_codegen/src/generator/rust/ty_primitive.rs
@@ -4,4 +4,8 @@ use crate::ir::*;
 #[derive(Debug, Clone)]
 pub struct TypePrimitiveGenerator(IrTypePrimitive);
 
-impl TypeRustGeneratorTrait for TypePrimitiveGenerator {}
+impl TypeRustGeneratorTrait for TypePrimitiveGenerator {
+    fn wire2api_body(&self) -> String {
+        "self".into()
+    }
+}
